pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mdp-bm-predictor"
version = "0.1.0"
description = "Local desktop GUI for MDP whole-body bone scintigraphy lesion prediction."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "numpy>=1.26,<3.0",
  "Pillow>=10,<13",
  "SimpleITK>=2.3,<3.0",
  "pyradiomics==3.0.1",
]

[project.scripts]
mdp-bm-predictor = "mdp_bm_predictor.__main__:main"

[tool.setuptools.packages.find]
include = ["mdp_bm_predictor*"]

[tool.setuptools.package-data]
mdp_bm_predictor = ["assets/*.json"]
